// web/src/lib/api/generated/models/githubComMantonxViewraInternalApplicationEnrichmentPipelineCircuitBreakerStatus.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */
import type { TimeDuration } from './timeDuration'
import type { GithubComMantonxViewraInternalApplicationEnrichmentPipelineCircuitState } from './githubComMantonxViewraInternalApplicationEnrichmentPipelineCircuitState'

export interface GithubComMantonxViewraInternalApplicationEnrichmentPipelineCircuitBreakerStatus {
  consecutive_failures?: number
  failure_threshold?: number
  last_failure?: string
  last_state_change?: string
  retry_after_seconds?: TimeDuration
  retry_at?: string
  stage?: string
  state?: GithubComMantonxViewraInternalApplicationEnrichmentPipelineCircuitState
}